win fix

+++ treedrawing/treedrawing.py
@@ -94,7 +94,7 @@ class Treedraw(object):
         tosave = tosave.replace("-FLAG", "")
         try:
             print "self.thefile is: ", self.thefile
-            if os.name == "nt":
+            if os.name == "nt" and os.path.isfile(self.thefile + '.bak'):                               
                 os.unlink(self.thefile + '.bak')
 
             os.rename(self.thefile, self.thefile + '.bak')
